Refactored NavigationRewritingExpressionVisitor.
- 'Flipped' the hierarchy to expose NavigationRewritingQueryModelVisitor instead
- Renamed NavigationJoin to NavigationClause (inheriting from IBodyClause, allowing them to be added to the query model at hand)
- Get rid of the _insideOrderByFlag by creating a specialized OrderByNavigationRewritingExpressionVisitor
- Get rid of the _navigationExpansionSubquery flag by performing explicit logic as necessary at the site of subquery construction
- Handle outer query references in subqueries by letting the outer query 'hoist' them out immediately after subquery visitation
- Handle deduplication of navigation clauses within VisitQueryModel
