+++ src/Microsoft.EntityFrameworkCore/Query/ExpressionVisitors/Internal/INavigationRewritingQueryModelVisitorFactory.cs
@@ -9,12 +9,12 @@ namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal
     ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
     ///     directly from your code. This API may change or be removed in future releases.
     /// </summary>
-    public interface INavigationRewritingExpressionVisitorFactory
+    public interface INavigationRewritingQueryModelVisitorFactory
     {
         /// <summary>
         ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
         ///     directly from your code. This API may change or be removed in future releases.
         /// </summary>
-        NavigationRewritingExpressionVisitor Create([NotNull] EntityQueryModelVisitor queryModelVisitor);
+        NavigationRewritingQueryModelVisitor Create([NotNull] EntityQueryModelVisitor queryModelVisitor);
     }
 }
