+++ src/Microsoft.EntityFrameworkCore/Query/ExpressionVisitors/Internal/NavigationRewritingQueryModelVisitorFactory.cs
@@ -1,10 +1,11 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using Microsoft.EntityFrameworkCore.Query.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
 using Microsoft.Extensions.DependencyInjection;
-using System;
+using JetBrains.Annotations;
 
 namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal
 {
@@ -12,7 +13,7 @@ namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal
     ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
     ///     directly from your code. This API may change or be removed in future releases.
     /// </summary>
-    public class NavigationRewritingExpressionVisitorFactory : INavigationRewritingExpressionVisitorFactory
+    public class NavigationRewritingQueryModelVisitorFactory : INavigationRewritingQueryModelVisitorFactory
     {
         private Lazy<IAsyncQueryProvider> _entityQueryProvider;
 
@@ -20,7 +21,7 @@ namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal
         ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
         ///     directly from your code. This API may change or be removed in future releases.
         /// </summary>
-        public NavigationRewritingExpressionVisitorFactory(IServiceProvider serviceProvider)
+        public NavigationRewritingQueryModelVisitorFactory([NotNull] IServiceProvider serviceProvider)
         {
             Check.NotNull(serviceProvider, nameof(serviceProvider));
 
@@ -32,7 +33,9 @@ namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal
         ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
         ///     directly from your code. This API may change or be removed in future releases.
         /// </summary>
-        public virtual NavigationRewritingExpressionVisitor Create(EntityQueryModelVisitor queryModelVisitor)
-            => new NavigationRewritingExpressionVisitor(queryModelVisitor, _entityQueryProvider.Value);
+        public virtual NavigationRewritingQueryModelVisitor Create(EntityQueryModelVisitor queryModelVisitor)
+        { 
+            return new NavigationRewritingQueryModelVisitor(queryModelVisitor, _entityQueryProvider.Value);
+        }
     }
 }
